combine wkt functions to one file
