--- sql/expression/function/point_from_text.go
@@ -1,123 +1,0 @@
-// Copyright 2020-2021 Dolthub, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package function
-
-import (
-	"fmt"
-	"strconv"
-	"strings"
-
-	"github.com/dolthub/go-mysql-server/sql/expression"
-
-	"github.com/dolthub/go-mysql-server/sql"
-)
-
-// PointFromText is a function that returns a point type from a WKT string
-type PointFromText struct {
-	expression.UnaryExpression
-}
-
-var _ sql.FunctionExpression = (*PointFromText)(nil)
-
-// NewPointFromText creates a new point expression.
-func NewPointFromText(e sql.Expression) sql.Expression {
-	return &PointFromText{expression.UnaryExpression{Child: e}}
-}
-
-// FunctionName implements sql.FunctionExpression
-func (p *PointFromText) FunctionName() string {
-	return "st_pointfromtext"
-}
-
-// Description implements sql.FunctionExpression
-func (p *PointFromText) Description() string {
-	return "returns a new point from a WKT string."
-}
-
-// IsNullable implements the sql.Expression interface.
-func (p *PointFromText) IsNullable() bool {
-	return p.Child.IsNullable()
-}
-
-// Type implements the sql.Expression interface.
-func (p *PointFromText) Type() sql.Type {
-	return p.Child.Type()
-}
-
-func (p *PointFromText) String() string {
-	return fmt.Sprintf("ST_POINTFROMTEXT(%s)", p.Child.String())
-}
-
-// WithChildren implements the Expression interface.
-func (p *PointFromText) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	if len(children) != 1 {
-		return nil, sql.ErrInvalidChildrenNumber.New(p, len(children), 1)
-	}
-	return NewPointFromText(children[0]), nil
-}
-
-// WKTToPoint expects a string like this "1.2 3.4"
-func WKTToPoint(s string) (sql.Point, error) {
-	// Empty string is wrong
-	if len(s) == 0 {
-		return sql.Point{}, sql.ErrInvalidGISData.New("ST_PointFromText")
-	}
-
-	// Get everything between spaces
-	args := strings.Fields(s)
-
-	// Check length
-	if len(args) != 2 {
-		return sql.Point{}, sql.ErrInvalidGISData.New("ST_PointFromText")
-	}
-
-	// Parse x
-	x, err := strconv.ParseFloat(args[0], 64)
-	if err != nil {
-		return sql.Point{}, sql.ErrInvalidGISData.New("ST_PointFromText")
-	}
-
-	// Parse y
-	y, err := strconv.ParseFloat(args[1], 64)
-	if err != nil {
-		return sql.Point{}, sql.ErrInvalidGISData.New("ST_PointFromText")
-	}
-
-	// Create point object
-	return sql.Point{X: x, Y: y}, nil
-}
-
-// Eval implements the sql.Expression interface.
-func (p *PointFromText) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	// Evaluate child
-	val, err := p.Child.Eval(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if val == nil {
-		return nil, nil
-	}
-
-	// Expect a string, throw error otherwise
-	if s, ok := val.(string); ok {
-		// TODO: possible to use a regular expression? "*point *\( *[0-9][0-9]* *[0-9][0-9]* *\) *" /gi
-		if geomType, data, err := ParseWKTHeader(s); err == nil && geomType == "point" {
-			return WKTToPoint(data)
-		}
-	}
-
-	return nil, sql.ErrInvalidGISData.New("ST_PointFromText")
-}
